feat: separate modules, rename SearchParams to Search, and move find_by fn's within Search
